requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
kafka-python-ng==2.2.2
redis==5.0.1
scikit-learn==1.4.0
numpy==1.26.3
joblib==1.3.2
